cli/python/requirements.txt
mysql-python
python-dotenv
